Move two components to common/
