+++ src/lib/common/ZoomOutMap.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
   import type { GeoJSON } from "geojson";
-  import icon from "../../assets/zoom_out_map.svg";
-  import { bbox } from "../maplibre_helpers";
-  import { map } from "../stores";
-  import { SecondaryButton } from "./govuk";
+  import icon from "../../../assets/zoom_out_map.svg";
+  import { bbox } from "../../maplibre_helpers";
+  import { map } from "../../stores";
+  import { SecondaryButton } from "../govuk";
 
   export let boundaryGeojson: GeoJSON;
